# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/device/code": {
      "post": {
        "description": "Returns an 8-character user code from an alphabet with no ambiguous glyphs, a 10-minute expiry and a 5-second polling interval. At most five flows may be in flight from one address.",
        "operationId": "DeviceController_request",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clientKind": {
                    "type": "string",
                    "enum": [
                      "desktop",
                      "bridge",
                      "premiere",
                      "ae",
                      "resolve"
                    ]
                  },
                  "hostApp": {
                    "type": "string",
                    "enum": [
                      "web",
                      "desktop",
                      "premiere",
                      "ae",
                      "resolve"
                    ]
                  },
                  "deviceInfo": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string",
                      "maxLength": 200
                    }
                  }
                },
                "required": [
                  "clientKind"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "A device code and a user code.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deviceCode": {
                      "type": "string"
                    },
                    "userCode": {
                      "type": "string"
                    },
                    "verificationUrl": {
                      "type": "string"
                    },
                    "verificationUrlComplete": {
                      "type": "string"
                    },
                    "interval": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "expiresIn": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "deviceCode",
                    "userCode",
                    "verificationUrl",
                    "verificationUrlComplete",
                    "interval",
                    "expiresIn"
                  ]
                }
              }
            }
          }
        },
        "summary": "Start a device sign-in",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/device/token": {
      "post": {
        "description": "400 with `auth/authorization_pending` until the code is approved, `auth/slow_down` when polled faster than the interval, `auth/expired_token` after 10 minutes and `auth/access_denied` when the user declines.",
        "operationId": "DeviceController_poll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "deviceCode": {
                    "type": "string",
                    "minLength": 16,
                    "maxLength": 256
                  }
                },
                "required": [
                  "deviceCode"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access and refresh tokens.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "tokenType": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "sessionId": {
                      "type": "string",
                      "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                    },
                    "workspaceId": {
                      "type": "string",
                      "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "owner",
                        "admin",
                        "editor",
                        "viewer"
                      ]
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "web",
                        "desktop",
                        "bridge",
                        "premiere",
                        "ae",
                        "resolve",
                        "api"
                      ]
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresIn",
                    "tokenType",
                    "sessionId",
                    "workspaceId",
                    "role",
                    "kind"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "auth/authorization_pending, auth/slow_down, auth/expired_token, auth/access_denied"
          }
        },
        "summary": "Poll for the result",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/device/code/{userCode}": {
      "get": {
        "description": "The approval screen's data: host application, device facts, address and a coarse location, so a user can tell a device on their desk from one on somebody else's.",
        "operationId": "DeviceController_describe",
        "parameters": [
          {
            "name": "userCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The pending request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userCode": {
                      "type": "string"
                    },
                    "clientKind": {
                      "type": "string",
                      "enum": [
                        "web",
                        "desktop",
                        "bridge",
                        "premiere",
                        "ae",
                        "resolve",
                        "api"
                      ]
                    },
                    "hostApp": {
                      "nullable": true,
                      "type": "string",
                      "enum": [
                        "web",
                        "desktop",
                        "premiere",
                        "ae",
                        "resolve"
                      ]
                    },
                    "deviceInfo": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "ip": {
                      "nullable": true,
                      "type": "string"
                    },
                    "location": {
                      "nullable": true,
                      "type": "object",
                      "properties": {
                        "country": {
                          "type": "string"
                        },
                        "region": {
                          "type": "string"
                        },
                        "city": {
                          "type": "string"
                        }
                      }
                    },
                    "expiresAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "userCode",
                    "clientKind",
                    "hostApp",
                    "deviceInfo",
                    "ip",
                    "location",
                    "expiresAt"
                  ]
                }
              }
            }
          }
        },
        "summary": "What is asking for approval",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/device/approve": {
      "post": {
        "description": "`workspaceId` defaults to the workspace of the approving session and is always re-checked against memberships.",
        "operationId": "DeviceController_decide",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCode": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 16
                  },
                  "workspaceId": {
                    "type": "string",
                    "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                  },
                  "decision": {
                    "type": "string",
                    "enum": [
                      "approve",
                      "deny"
                    ]
                  }
                },
                "required": [
                  "userCode"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "approved or denied"
          }
        },
        "summary": "Approve or decline a device sign-in",
        "tags": [
          "auth"
        ]
      }
    }
  },
  "info": {
    "title": "Aksharo API",
    "description": "Errors use the envelope `{ error: { code, message, details?, requestId } }` with `namespace/slug` codes (CONTRACTS section 8).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
...
  }
}